Creation de la task par XML

+++ unistall/unistall.ps1
@@ -1,6 +1,22 @@
 # Delete scheduled task
-schtasks /Delete /tn "JustinWPPrank" /F
+Write-Host "Deleting scheduled task..."
+
+# Check if scheduled task exists
+$task = schtasks /query /tn "JustinWPPrank"
+
+# If scheduled task exists, delete it
+if($task){
+    schtasks /delete /tn "JustinWPPrank" /f
+} else {
+    Write-Host "Scheduled task not found!"
+}
 
 # Delete install folder
+Write-Host "Deleting install folder..."
 $installDir = "$env:userprofile\jh"
-Remove-Item -Path $installDir -Recurse -Force
+if(Test-Path $installDir){
+    Remove-Item -Path $installDir -Recurse -Force
+    Write-Host "Done!"
+} else {
+    Write-Host "Install folder not found"
+}
